requirements.txt
python-pptx
Pillow
requests
pywin32; platform_system == "Windows"
